Refactor participant proof for Sonar gate
